backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib
bcrypt==3.2.2
fastapi-mail
python-dotenv
python-multipart
pydantic-settings
alembic
requests
jinja2
email-validator
sqladmin
itsdangerous
google-generativeai
fpdf2
pypdf
APScheduler==3.10.4
